refactor: 테스트 픽스처 추가 & 위도 경도 파라미터 위치 변경
