+++ backend/src/main/java/com/christmas/tree/domain/PointGenerator.java
@@ -13,27 +13,27 @@ import com.christmas.tree.exception.code.TreeErrorCode;
 public class PointGenerator {
 
     private static final int SRID = 4326;
-    private static final int MIN_LATITUDE = -180;
-    private static final int MAX_LATITUDE = 180;
-    private static final int MIN_LONGITUDE = -90;
-    private static final int MAX_LONGITUDE = 90;
+    private static final int MIN_LONGITUDE = -180;
+    private static final int MAX_LONGITUDE = 180;
+    private static final int MIN_LATITUDE = -90;
+    private static final int MAX_LATITUDE = 90;
 
     private PointGenerator() {}
 
-    public static Point generate(final Double latitude, final Double longitude) {
+    public static Point generate(final Double longitude, final Double latitude) {
         final GeometryFactory geometryFactory = new GeometryFactory(new PrecisionModel(), SRID);
-        if (latitude < MIN_LATITUDE || latitude > MAX_LATITUDE) {
-            throw new PointOutOfRangeException(
-                    TreeErrorCode.INVALID_LOCATION,
-                    Map.of("latitude", latitude.toString())
-            );
-        }
         if (longitude < MIN_LONGITUDE || longitude > MAX_LONGITUDE) {
             throw new PointOutOfRangeException(
                     TreeErrorCode.INVALID_LOCATION,
                     Map.of("longitude", longitude.toString())
             );
         }
-        return geometryFactory.createPoint(new Coordinate(latitude, longitude));
+        if (latitude < MIN_LATITUDE || latitude > MAX_LATITUDE) {
+            throw new PointOutOfRangeException(
+                    TreeErrorCode.INVALID_LOCATION,
+                    Map.of("latitude", latitude.toString())
+            );
+        }
+        return geometryFactory.createPoint(new Coordinate(longitude, latitude));
     }
 }
