+++ backend/src/test/java/com/christmas/acceptance/AcceptanceFixture.java
@@ -1,0 +1,29 @@
+package com.christmas.acceptance;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+
+import com.christmas.tree.repository.TreeRepository;
+
+import io.restassured.RestAssured;
+
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+public class AcceptanceFixture {
+
+    @LocalServerPort
+    private int port;
+
+    @BeforeEach
+    void setUp() {
+        RestAssured.port = port;
+    }
+
+    @AfterAll
+    static void tearDown(@Autowired TreeRepository treeRepository) {
+        treeRepository.deleteAll();
+    }
+}
